Add files via upload
Kaggle is the home of analytics and predictive modelling competitions. Data Science enthusiast, beginners to professionals, compete to create the
best predictive models using datasets uploaded both by individuals and companies looking for insights. Prizes can be as high as $3 million US. In late
2017 a survey of Kaggle users was conducted and received over 16,700 response.
The analysis and visualization has been done on the dataset to get the clearer idea about the data set